Migrated snapshot logic to new SnapshotOperationProvider
