+++ src/Runner.Worker/JobExtension.cs
@@ -409,24 +409,9 @@ namespace GitHub.Runner.Worker
                     var snapshotRequest = templateEvaluator.EvaluateJobSnapshotRequest(message.Snapshot, jobContext.ExpressionValues, jobContext.ExpressionFunctions);
                     if (snapshotRequest != null)
                     {
+                        var snapshotOperationProvider = HostContext.GetService<ISnapshotOperationProvider>();
                         jobContext.RegisterPostJobStep(new JobExtensionRunner(
-                            runAsync: async (IExecutionContext executionContext, object data) =>
-                            {
-                                var snapshotRequestFilePath = Path.Combine(HostContext.GetDirectory(WellKnownDirectory.Root), ".snapshot", "request.json");
-                                string snapshotRequestDirectoryPath = Path.GetDirectoryName(snapshotRequestFilePath);
-                                if (snapshotRequestDirectoryPath != null)
-                                {
-                                    Directory.CreateDirectory(snapshotRequestDirectoryPath);
-                                }
-
-                                var snapshotRequestJson = JsonConvert.SerializeObject(snapshotRequest);
-                                await File.WriteAllTextAsync(snapshotRequestFilePath, snapshotRequestJson);
-                                executionContext.Output($"A snapshot request was created with parameters: {snapshotRequestJson}");
-                                executionContext.Output($"Request written to: {snapshotRequestFilePath}");
-                                executionContext.Output("This request will be processed after the job completes. You will not receive any feedback on the snapshot process within the workflow logs of this job.");
-                                executionContext.Output("If the snapshot process is successful, you should see a new image with the requested name in the list of available custom images when creating a new GitHub-hosted Runner.");
-                                await Task.CompletedTask;
-                            }, 
+                            runAsync: (executionContext, _) => snapshotOperationProvider.CreateSnapshotRequestAsync(executionContext, snapshotRequest), 
                             condition: $"{PipelineTemplateConstants.Success}()",
                             displayName: $"Create custom image",
                             data: null));
